requirements.txt
torch>=2.0
transformers>=4.40
datasets>=2.18
huggingface-hub>=0.23
tokenizers>=0.19
accelerate>=0.30
gguf>=0.10.0
pyyaml>=6.0
numpy>=1.24
tqdm>=4.66
